Create 13 benefits of walnuts,
There is no doubt that dry fruits are beneficial for health. Along with a healthy diet, if you include dryfruits in your routine, it is good for your health. Along with almonds, raisins, dates and walnuts, there are many dry fruits that help to maintain your health. At the moment we talk only of walnuts, which is a treasure of many qualities.

Walnuts are called Walnut in English, Akruta Kaya in Telugu, Akrothandi in Malayalam, Akrota in Kannada, Akrottu in Tamil, Akarot in Marathi and Akrot in Gujarati. The names of walnuts in different languages ​​have the same benefits. In this article of Stylecrase, we will explain the benefits of eating walnuts.